Mention how to discover existing sbt keys
The introduction to build definition goes in detail about what keys are and how to create them, but there's no information about what the list of keys are, and the user is left to pick up from the code samples that `name` or `organization` seem to exist. 

I suggest to add this short set of paragraph to let the user know how they can discover the current settings and tasks.